+++ packages/hms-video-store/src/sdk/store/StoreInterfaces.ts
